docs: set `site_url` to fix broken 404 page on RTD (#450)
### Summary of Changes

The 404 page was previously not properly rendered on Read the Docs
(RTD). This PR overrides the build commands of RTD. While this makes the
404 page work properly, it also removes the RTD flyout to select a
version. For now, however, we deem the 404 page as more essential.

Once https://github.com/readthedocs/readthedocs.org/issues/8529 is done,
we can use the default build again to make everything work as expected.

---------

Co-authored-by: megalinter-bot <129584137+megalinter-bot@users.noreply.github.com>