feat: remove type arguments from calls (#581)
### Summary of Changes

Calls no longer have a type argument list. Since all objects we expose
are assumed to be immutable, there is no reason to specify the type
arguments of a call. We can always infer them from the parameters or set
them to `Any`/`Nothing` if the parameters don't provide enough
information. An added benefit of this is that our grammar now only
requires a finite lookahead.