feat: list & map literals (#619)
Closes #615
Closes #616

### Summary of Changes

Add list and map literals. The computation of the element & key/value
types will happen in a later pull request, together with type
computation of other type parameters
(https://github.com/Safe-DS/DSL/issues/23).

This PR also removes `vararg` parameters in favor of a parameter with
list types. This simplifies the graphical view and gets rid of a bunch
of validation rules.

---------

Co-authored-by: megalinter-bot <129584137+megalinter-bot@users.noreply.github.com>