refactor: better method names (#661)
### Summary of Changes

I've dropped the lengthy `OrUndefined` and `OrEmpty` suffixes from
methods of the `NodeMapper` and the functions to access node properties.
Instead, they now have a `stream` prefix if they return a stream or a
`get` prefix/no prefix otherwise. This better aligns with the name
convention of API elements provided by `langium`.