feat: stop validation after lexing/parsing errors (#662)
### Summary of Changes

Linking or custom validation errors are no longer displayed if a
document has lexing or parsing errors. This better highlights the
criticality of those issues and ensures that users fix those first. It
also prevents the display of some additional errors that were caused by
the lexing or parsing errors.