feat: error if type parameters don't have sufficient context (#687)
### Summary of Changes

Show an error if we already know at the declaration-site that we won't
be able to infer type parameters in a call. A type parameter must occur
in the parameter list of the containing callable. However, it does not
suffice if they appear inside a union type or the parameter list of a
callable type.