feat: stricter definition of `const` parameters (#776)
### Summary of Changes

The definition of `const` parameters is now more strict:
* Previously, any value that could be fully evaluated by the partial
evaluator could be assigned to a constant parameter.
* Now, a whitelist of expressions is allowed, including
    * boolean/float/int/null/string literals,
    * list literals (where the elements are constant),
    * map literals (where keys and values are constant),
    * enum variants (where arguments are constant).

This allows such parameters to be handled fully in the sidebar of the
graphical view. It also makes it possible to show errors/warnings at the
exact places they occur, like when checking whether an
`AnnotationTarget` appears multiple times.