fix: handling of default values in entry point of call graph computation (#797)
### Summary of Changes

Consider the following code:

```
@Pure fun default() -> r: Any

segment mySegment(param: () -> () = default) {
    »param()«;
}
```

When computing the call graph for the `param()` call, we were previously
always substituting `param` with its default value, since the parameter
substitutions were empty in the entry point. Because of this, we were
incorrectly considering the `param()` call as pure, just like the
segment `mySegment`.