feat: error if type parameter has multiple lower/upper bounds (#870)
Closes #867

### Summary of Changes

Just like we don't have multiple inheritance or intersection types, we
only want to deal with a single lower and a single upper bound for type
parameters. We currently don't have a use-case for multiple bounds, so
we may as well simplify the implementation of the DSL. This PR adds a
check to ensure that type parameters have a single lower and a single
upper bound.