feat: simplify union types containing type parameter types disregarding entry order (#894)
### Summary of Changes

Consider the `class C<T> where { T sub Any }`. Previously, the type
`union<T, Any>` was simplified to `T`, while the type `union<Any, T>`
was simplified to `Any`. Now the simplified type is `T` in both cases.
We chose the type parameter type, since it can be narrower if it has a
lower bound.