feat: handle invariant/covariant type parameters when computing lowest common supertype (#868)
Closes partially #860

### Summary of Changes

Invariant/covariant type parameters are now handled when computing the
lowest common supertype of a list of types. This functionality is used
to compute the type of
* lists (based on their elements)
* maps (based on their keys and values)
* elvis operator (based on their operands).

It's also needed later for #861. Contravariant type parameters need a
means to compute the highest common subtype, which will be tackled in a
future PR.