refactor: add type factory service
This way we can keep the interface to create types basically unchanged, but we can inject services into the instances. This is a) helpful, since types would now be able to fully simplify themselves, and b) needed because setting the nullability of an empty intersection type to false yields `Any`. Previously, we had no way to express this.