fix: simplification of union types (#897)
### Summary of Changes

Consider the following program:

```
class C<T> {
    attr a: union<Int, T>
}

segment s(p: C<String>) {
    val v = p.a;
}
```

Previously, the type of the placeholder `v` was inferred to be `String`,
because we incorrectly simplified union types that contained entries
that were not fully substituted. Here, we replaced the entry `Int` with
`T`, since `Int` is a subtype of `T`. This can change, however, once we
substitute `T`, as shown here.

Now, we never consider types that are not fully substituted for
replacement and compute the correct type `union<Int, String>`.