feat: compute highest common subtype (#901)
Closes #860

### Summary of Changes

Add a function to compute the highest common subtype of a list of types.
This is needed to compute the lowest common supertype of a list of types
with contravariant type parameters. Example:

```
class Consumer<in T>
class Logger sub Consumer<String>

segment mySegment(
    p1: Consumer<Any>,
    p2: Logger,
) {
    val a = [p1, p2];
}
```

We can now infer the type of the placeholder `a` to
`List<Consumer<String>>`, where previously we would have inferred
`List<Consumer<Nothing>>`.