feat: do not check whether optional type parameters of classes can be inferred (#1090)
Closes #1084

### Summary of Changes

Optional type parameters of classes are now always allowed, even if they
can never be inferred from their usage in the constructor.

This is useful to initialize them to some default value upon
instantiation and override this value later, e.g. in return types of
functions. See #1084 for an example.