feat: eda plot view (#1161)
Closes #955
Closes #986

### Summary of Changes

Implemented history/action structure for webview to add new actions to
state and have them (if external and info not already present) send
execute requests to runner that are then cancellable or deployed in
correct order. Only fully working for Plots/Tabs at the moment, that are
on deploy added to tabs state and set as currentIndex. All Tabs and
Table retain their state. Runner uses existing methods in RunnerAPI to
get back to relevant state by executing past manipulating actions and
then returns the result of the new action (only if plots right now).

Tabs can be created by selecting columns and right clicking, zooming in
on profiling images or by creating an empty tab with the plus icon in
the sidebar. There in the guided menu users can change the current Tab
to display other info. At that point the tab will go into building state
and show prompts, loading screens and buttons accordingly. Typings are
adapted to abstract as much as possible (mainly around column count for
tabs, none, one and two) and stores are heavily used for reactivity.

---------

Co-authored-by: Lars Reimann <mail@larsreimann.com>
Co-authored-by: WinPlay02 <winplay02_gh@woberlaender.de>
Co-authored-by: megalinter-bot <129584137+megalinter-bot@users.noreply.github.com>