docs: first regression program (#1202)
Closes partially #1031

### Summary of Changes

This pull request adds a new page to the getting started tab. This page
displays the first steps one has to follow, in order to write their
first safe ds program, that implements a regression model. It also
offers a section that shows the full code, so that one can copy and
paste for convenience.

---------

Co-authored-by: thelenf0 <s1frthel@uni-bonn.de>
Co-authored-by: Lars Reimann <mail@larsreimann.com>