docs: exclude stub/implementation code blocks from search (#1258)
### Summary of Changes

When searching for names of API elements, the search previously only
showed the stub/implementation code blocks. This was confusing and
unhelpful, since users would rather see a description of the API
element.

This PR, hence, excludes the code blocks from search, which also reduces
the size of the search index and speeds up the search as a side effect.